+++ apps/ledger-live-mobile/src/families/elrond/accountActions.tsx
@@ -82,7 +82,7 @@ const getMainActions = (props: getActionsType): getActionsReturnType => {
       event: "button_clicked",
       eventProperties: {
         button: "stake",
-        token: "ELROND",
+        currency: "ELROND",
         page: "Account Page",
       },
     },