+++ libs/ledger-live-common/src/families/bitcoin/wallet-btc/explorer/index.ts
@@ -6,7 +6,7 @@ import { Address, Block, TX } from "../storage/types";
 import { IExplorer } from "./types";
 import { errorInterceptor } from "../../../../network";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
-import { blockchainBaseURL } from "../../../../api/Ledger";
+import { blockchainBaseURL } from "../../../../explorer";
 
 type ExplorerParams = {
   batch_size?: number;